+++ backend_springboot/src/main/java/com/MyourCar/domain/user/UserRepository.java
@@ -1,0 +1,6 @@
+package com.MyourCar.domain.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
