--- backend_springboot/src/main/java/com/MyourCar/domain/users/UsersRepository.java
@@ -1,7 +1,0 @@
-package com.MyourCar.domain.users;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-public interface UsersRepository extends JpaRepository<Users, Long> {
-}
