+++ backend_springboot/src/main/java/com/MyourCar/service/user/UserService.java
@@ -1,0 +1,21 @@
+package com.MyourCar.service.user;
+
+import com.MyourCar.domain.user.User;
+import com.MyourCar.domain.user.UserRepository;
+import com.MyourCar.web.dto.UserResponseDto;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class UserService {
+    private final UserRepository usersRepository;
+
+    public UserResponseDto findById(Long id) {
+        User entity = usersRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("해당 유저가 없습니다. id=" + id));
+
+        return new UserResponseDto(entity);
+    }
+
+}
