+++ backend_springboot/src/test/java/com/MyourCar/web/dto/UserResponseDtoTest.java
@@ -4,6 +4,6 @@ import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class UsersResponseDtoTest {
+public class UserResponseDtoTest {
 
 }
